refactor: rename AutoCompleteCountryNew to AutoCompleteCountry
- also removes old AutoCompleteCountry component (unused)
